REST
http://localhost:8090/getOrder?name=username&pass=password&item=flower-4000

Can have as many items as you want.

+++ PhilliesFlowersBack/src/main/java/com/phillies/domain/OrderItem.java
@@ -29,7 +29,7 @@ public class OrderItem {
 		this.quantity = quantity;
 	}
 	
-	public float getCost() {
+	public double getCost() {
 		return this.item.getPrice() * this.quantity;
 	}
